Add files via upload
File Watcher watches a folder or set of folders for new or updated files. When it finds one, it passes the filename to a script that you designate - to perform whatever action you desire. File Watcher allows all user settings to be saved and opened so that the same tasks can be repeated in the future without re-keying.